requirements.txt
aiohttp==3.11.13
async-timeout==4.0.3
lz4==4.4.4
numpy==2.3.5
orjson==3.10.18
pandas==2.2.3
matplotlib==3.9.2
pydantic==2.11.5
pydantic_settings==2.9.1
PyYAML==6.0.2
python-dotenv==1.1.0
redis==4.5.5
rich==14.0.0
scipy==1.15.3
tenacity==9.1.2
websockets==15.0
